+++ NEWS.md
@@ -1,0 +1,5 @@
+# 0.0.1
+
+## Features
+
+* Validate your Rails response JSON with `match_response_schema`
