+++ json-matchers.gemspec
@@ -1,16 +1,16 @@
 # coding: utf-8
 lib = File.expand_path("../lib", __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require "json/schema/matchers/version"
+require "json/matchers/version"
 
 Gem::Specification.new do |spec|
-  spec.name          = "json-schema-matchers"
-  spec.version       = JSON::Schema::Matchers::VERSION
+  spec.name          = "json-matchers"
+  spec.version       = JSON::Matchers::VERSION
   spec.authors       = ["Sean Doyle"]
   spec.email         = ["seandoyle@thoughtbot.com"]
   spec.summary       = %q{Validate your Rails JSON API's JSON}
   spec.description   = %q{Validate your Rails JSON API's JSON}
-  spec.homepage      = "https://github.com/seanpdoyle/json-schema-matchers"
+  spec.homepage      = "https://github.com/seanpdoyle/json-matchers"
   spec.license       = "MIT"
 
   spec.files         = `git ls-files -z`.split("\x0")
