+++ lib/json/matchers/errors.rb
@@ -1,0 +1,7 @@
+module JSON
+  module Matchers
+    InvalidError = Class.new(StandardError)
+    DoesNotMatch = Class.new(InvalidError)
+    MissingSchema = Class.new(Errno::ENOENT)
+  end
+end
