+++ lib/json/matchers/rspec.rb
@@ -1,5 +1,5 @@
 if RSpec.respond_to?(:configure)
   RSpec.configure do |config|
-    config.include JSON::Schema::Matchers
+    config.include JSON::Matchers
   end
 end
