+++ lib/json/matchers/version.rb
@@ -1,0 +1,5 @@
+module JSON
+  module Matchers
+    VERSION = "0.0.1"
+  end
+end
