--- lib/json/schema/matchers/matcher.rb
@@ -1,23 +1,0 @@
-require "json-schema"
-
-module JSON
-  class Schema
-    module Matchers
-      Matcher = Struct.new(:schema) do
-        def matches?(response)
-          JSON::Validator.validate!(json_schema, response.body, strict: true)
-        rescue JSON::Schema::ValidationError
-          raise DoesNotMatch, response.body
-        rescue JSON::ParserError
-          raise InvalidError
-        end
-
-        private
-
-        def json_schema
-          JSON.parse(schema)
-        end
-      end
-    end
-  end
-end
