--- lib/json/schema/matchers/version.rb
@@ -1,7 +1,0 @@
-module JSON
-  class Schema
-    module Matchers
-      VERSION = "0.0.1"
-    end
-  end
-end
