+++ spec/json/matchers/schema_parser_spec.rb
@@ -1,11 +1,11 @@
-require "json/schema/matchers/schema_parser"
+require "json/matchers/schema_parser"
 
-describe JSON::Schema::Matchers::SchemaParser do
+describe JSON::Matchers::SchemaParser do
   describe "#schema_for" do
     it "returns the JSON string for a given schema" do
       create_schema("foo", { type: "array" })
 
-      schema_parser = JSON::Schema::Matchers::SchemaParser.new(schema_root)
+      schema_parser = JSON::Matchers::SchemaParser.new(schema_root)
 
       expect(schema_parser.schema_for("foo")).to eq '{"type":"array"}'
     end
@@ -35,7 +35,7 @@ describe JSON::Schema::Matchers::SchemaParser do
       }
       JSON
 
-      schema_parser = JSON::Schema::Matchers::SchemaParser.new(schema_root)
+      schema_parser = JSON::Matchers::SchemaParser.new(schema_root)
       schema = schema_parser.schema_for("posts")
 
       expect(JSON.parse(schema)).to eq({
@@ -59,10 +59,10 @@ describe JSON::Schema::Matchers::SchemaParser do
     end
 
     it "fails when the schema is missing" do
-      schema_parser = JSON::Schema::Matchers::SchemaParser.new(schema_root)
+      schema_parser = JSON::Matchers::SchemaParser.new(schema_root)
       expect {
         schema_parser.schema_for("missing")
-      }.to raise_error(JSON::Schema::Matchers::MissingSchema, /missing.json/)
+      }.to raise_error(JSON::Matchers::MissingSchema, /missing.json/)
     end
   end
 end
